backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohappyeyeballs==2.4.3",
    "aiohttp==3.10.8",
    "aioredis==1.3.1",
    "aiosignal==1.3.1",
    "amqp==5.2.0",
    "annotated-types==0.7.0",
    "anyio==4.6.0",
    "asgiref==3.8.1",
    "async-timeout==4.0.3",
    "attrs==24.2.0",
    "autobahn==24.4.2",
    "automat==24.8.1",
    "autopep8==2.3.1",
    "beautifulsoup4==4.12.3",
    "billiard==4.2.1",
    "cachetools==5.5.0",
    "celery==5.4.0",
    "certifi==2024.7.4",
    "cffi==1.17.0",
    "channels-redis==3.4.1",
    "channels==3.0.5",
    "charset-normalizer==3.3.2",
    "click-didyoumean==0.3.1",
    "click-plugins==1.1.1",
    "click-repl==0.3.0",
    "click==8.1.7",
    "colorama==0.4.6",
    "constantly==23.10.4",
    "cron-descriptor==1.4.5",
    "cryptography==43.0.0",
    "daphne==3.0.2",
    "defusedxml==0.7.1",
    "dj-rest-auth==6.0.0",
    "django-allauth==0.57.1",
    "django-celery-beat==2.7.0",
    "django-celery-results==2.5.1",
    "django-cors-headers==4.4.0",
    "django-resized==1.0.3",
    "django-silk==5.2.0",
    "django-timezone-field==7.0",
    "django==5.1",
    "djangorestframework-simplejwt==5.3.1",
    "djangorestframework==3.15.2",
    "frozenlist==1.4.1",
    "google-ai-generativelanguage==0.6.6",
    "google-api-core==2.19.2",
    "google-api-python-client==2.144.0",
    "google-auth-httplib2==0.2.0",
    "google-auth==2.34.0",
    "google-generativeai==0.7.2",
    "googleapis-common-protos==1.65.0",
    "gprof2dot==2024.6.6",
    "greenlet==3.1.1",
    "grpcio-status==1.62.3",
    "grpcio==1.66.1",
    "h11==0.14.0",
    "hiredis==3.0.0",
    "httpcore==1.0.5",
    "httplib2==0.22.0",
    "httpx==0.27.2",
    "hyperlink==21.0.0",
    "idna==3.8",
    "incremental==24.7.2",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "kombu==5.4.2",
    "langchain-core==0.3.7",
    "langchain-google-genai==2.0.0",
    "langchain-text-splitters==0.3.0",
    "langchain==0.3.1",
    "langsmith==0.1.129",
    "msgpack==1.1.0",
    "multidict==6.1.0",
    "numpy==1.26.4",
    "oauthlib==3.2.2",
    "orjson==3.10.7",
    "packaging==24.1",
    "pillow==10.4.0",
    "prompt-toolkit==3.0.48",
    "proto-plus==1.24.0",
    "protobuf==4.25.4",
    "pyasn1-modules==0.4.0",
    "pyasn1==0.6.0",
    "pycodestyle==2.12.1",
    "pycparser==2.22",
    "pydantic-core==2.23.3",
    "pydantic==2.9.1",
    "pyjwt==2.9.0",
    "pyopenssl==24.2.1",
    "pyparsing==3.1.4",
    "python-crontab==3.2.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "python-slugify==8.0.4",
    "python3-openid==3.2.0",
    "pyyaml==6.0.2",
    "razorpay==1.4.2",
    "redis==5.0.8",
    "requests-oauthlib==2.0.0",
    "requests==2.32.3",
    "rsa==4.9",
    "service-identity==24.1.0",
    "six==1.16.0",
    "slugify==0.0.1",
    "sniffio==1.3.1",
    "soupsieve==2.6",
    "sqlalchemy==2.0.35",
    "sqlparse==0.5.1",
    "tenacity==8.5.0",
    "text-unidecode==1.3",
    "tqdm==4.66.5",
    "twisted==24.7.0",
    "txaio==23.1.1",
    "typing-extensions==4.12.2",
    "tzdata==2024.1",
    "uritemplate==4.1.1",
    "urllib3==2.2.2",
    "vine==5.1.0",
    "wcwidth==0.2.13",
    "yarl==1.13.1",
    "zope-interface==7.0.3",
]
